implement search
With this commit we will be able to search for any feature or job class
and see their code when found. The search is fuzzy using 30% match of
the word so it is very likely to have results with very little match.

The code preview component has been implemented as a separate
“component” (not a vue component but a blade component) and provides
the functionality with a mixin (CodePreview).